llm-safety-framework-public/requirements.txt
# LLM Safety Framework - Requirements
# ====================================

# Core
pydantic>=2.0
pyyaml>=6.0
httpx>=0.25
aiofiles>=23.0

# LLM Providers
openai>=1.0
anthropic>=0.18
mistralai>=0.1
together>=1.0

# API Server
fastapi>=0.109
uvicorn>=0.27
python-multipart>=0.0.6

# Data Processing
pandas>=2.0
numpy>=1.24

# CLI
typer[all]>=0.9
rich>=13.0

# Database
sqlalchemy>=2.0
aiosqlite>=0.19

# Utilities
python-dotenv>=1.0
tenacity>=8.2

# Development (install with: pip install -r requirements-dev.txt)
# pytest>=7.4
# pytest-asyncio>=0.21
# pytest-cov>=4.1
# black>=23.0
# ruff>=0.1
# mypy>=1.5
